# go/cash-register-service-grpc/shared/api/checkout/checkout.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "register checkout service",
    "version": "version not set"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/health": {
      "get": {
        "operationId": "Health",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/checkoutHealthResponse"
            }
          }
        },
        "tags": [
          "CheckoutService"
        ]
      }
    },
    "/v1/checkout/items": {
      "post": {
        "summary": "catalog logically belongs to a separate service\nmoved it here due to limited time",
        "operationId": "AddItem",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/checkoutAddItemResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/checkoutAddItemRequest"
            }
          }
        ],
        "tags": [
          "CheckoutService"
        ]
      }
    },
    "/v1/checkout/receipt": {
      "post": {
        "operationId": "Checkout",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/checkoutCheckoutResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/checkoutCheckoutRequest"
            }
          }
        ],
        "tags": [
          "CheckoutService"
        ]
      }
    }
  },
  "definitions": {
    "checkoutAddItemRequest": {
      "type": "object",
      "properties": {
        "item": {
          "$ref": "#/definitions/checkoutItem"
        }
      }
    },
    "checkoutAddItemResponse": {
      "type": "object",
      "properties": {
        "item": {
          "$ref": "#/definitions/checkoutItem"
        }
      }
    },
    "checkoutCheckoutRequest": {
      "type": "object",
      "properties": {
        "register": {
          "type": "string"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/checkoutScanItem"
          }
        }
      }
    },
    "checkoutCheckoutResponse": {
      "type": "object",
      "properties": {
        "receipt": {
          "$ref": "#/definitions/checkoutReceipt"
        }
      },
      "title": "end checkout  will return the receipt"
    },
    "checkoutHealthResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "time": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "Address": {
          "type": "string"
        },
        "version": {
          "type": "string"
        },
        "build": {
          "type": "string"
        }
      },
      "title": "field numbers 1 through 15 take one byte to encode"
    },
    "checkoutItem": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "price": {
          "type": "number",
          "format": "float"
        },
        "valid": {
          "type": "boolean",
          "format": "boolean"
        }
      }
    },
    "checkoutLineItem": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "qty": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        },
        "price": {
          "type": "number",
          "format": "float"
        },
        "extprice": {
          "type": "number",
          "format": "float"
        }
      }
    },
    "checkoutReceipt": {
      "type": "object",
      "properties": {
        "register": {
          "type": "string"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/checkoutLineItem"
          }
        },
        "time": {
          "type": "string",
          "format": "int64"
        },
        "total": {
          "type": "number",
          "format": "float"
        }
      }
    },
    "checkoutScanItem": {
      "type": "object",
      "properties": {
        "upc": {
          "type": "string"
        },
        "time": {
          "type": "string",
          "format": "int64"
        }
      }
    }
  }
}